+++ json_serializable/test/kitchen_sink/kitchen_sink.g_exclude_null__no_encode_empty__use_wrappers.g.dart
@@ -310,25 +310,35 @@ class _$JsonConverterTestClassJsonMapWrapper extends $JsonMapWrapper {
 
   @override
   Iterable<String> get keys sync* {
-    if (_v.duration != null) {
+    if ((_v.duration == null ? null : durationConverter.toJson(_v.duration)) !=
+        null) {
       yield 'duration';
     }
     if (_v.durationList?.isNotEmpty ?? false) {
       yield 'durationList';
     }
-    if (_v.bigInt != null) {
+    if ((_v.bigInt == null
+            ? null
+            : const BigIntStringConverter().toJson(_v.bigInt)) !=
+        null) {
       yield 'bigInt';
     }
     if (_v.bigIntMap?.isNotEmpty ?? false) {
       yield 'bigIntMap';
     }
-    if (_v.numberSilly != null) {
+    if ((_v.numberSilly == null
+            ? null
+            : TrivialNumberConverter.instance.toJson(_v.numberSilly)) !=
+        null) {
       yield 'numberSilly';
     }
     if (_v.numberSillySet?.isNotEmpty ?? false) {
       yield 'numberSillySet';
     }
-    if (_v.dateTime != null) {
+    if ((_v.dateTime == null
+            ? null
+            : const EpochDateTimeConverter().toJson(_v.dateTime)) !=
+        null) {
       yield 'dateTime';
     }
   }
@@ -403,7 +413,8 @@ class _$JsonConverterGenericJsonMapWrapper<S, T, U> extends $JsonMapWrapper {
 
   @override
   Iterable<String> get keys sync* {
-    if (_v.item != null) {
+    if ((_v.item == null ? null : GenericConverter<S>().toJson(_v.item)) !=
+        null) {
       yield 'item';
     }
     if (_v.itemList?.isNotEmpty ?? false) {